+++ packages/ast-utilities/CHANGELOG.md
@@ -9,7 +9,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
-- `addImportSpecifier()` doesn't allow duplicate specifiers
+- `addImportSpecifier()` doesn't allow duplicate specifiers [[#2081)](https://github.com/Shopify/quilt/pull/2081)]
 
 ## 1.1.2 - 2021-11-23